pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hermes-dashboard"
version = "0.1.0"
description = "Minimal read-only web dashboard for Hermes Agent"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "starlette>=0.36.0",
    "jinja2>=3.1.0",
    "uvicorn[standard]>=0.24.0",
    "pyyaml>=6.0",
    "markdown>=3.5",
]

[project.scripts]
hermes-dashboard = "dashboard:main"

[tool.hatch.build.targets.wheel]
packages = ["dashboard"]
